Update Final Project.py
I just did a copy/paste of the whole program as is updated using checkboxes instead of radiobuttons. Either way is fine with me. I also added check clears at the bottom for comments, emails, and the reply for how they heard about us as well as getting rid of the comment lines for adding a logo image.